feat: remove aic.fish function and update gh-clone-all.tmpl and aur.txt.tmpl
- Remove the `aic.fish` function from `arch/dot_config/fish/pkg/aicommits/functions/aic.fish`.
- Update `arch/dot_local/bin/executable_gh-clone-all.tmpl` to pull changes from the remote repository if the destination path already exists.
- Update `arch/dot_local/bin/executable_gh-clone-all.tmpl` to create the parent directories before cloning the repository using `gh repo clone`.
- Update `arch/other/pkg/aur.txt.tmpl` to replace `gnome-shell-extension-system-monitor-next-git` with `gnome-shell-extension-vitals`.
- Update `install.sh` to pass command line arguments to `chezmoi apply`.

This commit removes the `aic.fish` function as it is no longer needed. It also updates the `gh-clone-all.tmpl` script to pull changes from the remote repository if the destination path already exists. Additionally, it adds the functionality to create the parent directories before cloning the repository using `gh repo clone`. The `aur.txt.tmpl` file is updated to replace the package `gnome-shell-extension-system-monitor-next-git` with `gnome-shell-extension-vitals`. Finally, the `install.sh` script is modified to pass command line arguments to `chezmoi apply`.