feat(chezmois): add support for executable files in chezmoi setup
This commit adds support for executable files in the chezmoi setup process. Previously, all files were installed with read and write permissions for the owner and read permissions for the group and others. Now, if a file has a name starting with "executable_", it will be installed with execute permissions for the owner, group, and others.

The following files have been modified to support this feature:
- `arch/.chezmoiscripts/run_before_setup.sh.tmpl`
- `arch/other/root/etc/cron.daily/executable_trash-empty.sh.tmpl`

Additionally, the following files have been added as executable files:
- `arch/other/root/etc/cron.daily/executable_geoip.sh`
- `arch/other/root/etc/cron.daily/executable_geosite.sh`
- `arch/other/root/etc/cron.weekly/executable_sing-box-ui.sh`

These changes improve the flexibility and usability of the chezmoi setup process.