feat(direnvrc): add support for layout_conda and layout_micromamba functions
functions: `layout_conda` and `layout_micromamba`.

The `layout_conda` function checks for the presence of Conda in the system and activates the specified environment. If no environment is specified, it uses the name of the current directory as the environment name. If the environment does not exist, it creates it using the `conda create` command.

The `layout_micromamba` function activates the specified environment using Micromamba. If no environment is specified, it uses the name of the current directory as the environment name.

These changes were made to enhance the functionality of the `direnvrc` file and provide support for managing Conda and Micromamba environments.