feat(arch): update run_after_gnome.sh.tmpl
- Set `org.gnome.shell.extensions.vitals` `fixed-widths` to `false`
- Set `org.gnome.shell.extensions.vitals` `hot-sensors` to `["_processor_usage_", "_memory_usage_", "__network-rx_max__", "__temperature_max__"]`

These changes are made to update the configuration of the `run_after_gnome.sh.tmpl` script in the `arch` directory. The `fixed-widths` option is set to `false` to allow dynamic widths for the `org.gnome.shell.extensions.vitals` extension. The `hot-sensors` option is updated to include additional sensors for monitoring processor usage, memory usage, network receive maximum, and temperature maximum.