feat(systemd): add script to run after DDNS update
This commit adds a new script `run_after_ddns.sh` to the `systemd/user` directory. The script sets the necessary options for error handling and then calls the `ddns install` command. This script will be executed after a DDNS update to perform additional actions.

The purpose of this script is to automate the execution of certain tasks that need to be performed after a DDNS update. By adding this script to the `systemd/user` directory, it can be easily integrated into the system's startup process and ensure that the necessary actions are taken after a DDNS update.