fix(arch): update file URLs and command names
Update the URLs of files in the `.chezmoiexternal.toml` file to point to the correct locations.
Change the command names in the `run_after_ddns.sh` and `run_after_claps.fish` files to reflect the updated file URLs.

These changes are made to ensure that the correct files are being accessed and executed as intended.