feat(srv): add Forgejo service
This commit introduces a self-hosted Forgejo service, managed via Docker Compose. It includes the necessary configuration for the service, a setup script to manage the user and permissions, and integrates it into the `cachyos` profile.

The Forgejo service is configured to use a dedicated user and exposes both SSH and web UI ports.

Additionally, this commit disables IPv6 networking for the Jellyfin service as a workaround, because it was causing the service to fail to start correctly.